refactor(edit-goal): extract EditGoalStepList from EditGoalView
Cut the step-row list layer — rows, sub-step blocks, add-step affordance,
inline-rename/evidence-picker/delete-confirm state, and the two modals those
drive — out of EditGoalView into a new sibling EditGoalStepList (issue #489).
EditGoalView is now a thin composition; shared types stay defined there (D2).
Zero behaviour change: EditGoalView.test.tsx passes unmodified.
