+++ apps/native-rd/src/components/EditGoalView/EditGoalView.tsx
@@ -17,34 +17,26 @@
  * be verified before the screen is wired.
  * `grep -rn "EditGoalView" src/screens` stays empty until then.
  *
+ * This component is a thin composition (header, goal-title card, optional
+ * description, EditGoalStepList, dates info banner, Done footer). The step-row
+ * list layer — rows, sub-step blocks, add-step affordance, inline-rename /
+ * evidence-picker / delete-confirm state, and the two modals those drive —
+ * lives in EditGoalStepList (issue #489), which the New Goal wizard reuses
+ * (#490). The shared step/sub-step types stay defined here (D2).
+ *
  * Drag orchestration lives in useEditGoalDrag; the row anatomy in
  * EditGoalStepRow; the ⋯ menu content in EditGoalOverflowMenu.
  */
-import React, { useEffect, useState } from "react";
-import {
-  View,
-  Text as RNText,
-  TextInput,
-  Pressable,
-  Modal,
-  AccessibilityInfo,
-} from "react-native";
-import { SafeAreaView } from "react-native-safe-area-context";
-import { GestureHandlerRootView } from "react-native-gesture-handler";
+import React from "react";
+import { View, Text as RNText, TextInput } from "react-native";
 import { DotsThree, Pencil } from "phosphor-react-native";
 import { useUnistyles } from "react-native-unistyles";
-import { useAnimationPref } from "../../hooks/useAnimationPref";
-import { Text } from "../Text";
 import { IconButton } from "../IconButton";
 import { Button } from "../Button";
 import { ScreenSubHeader } from "../ScreenHeader/ScreenSubHeader";
-import { EvidenceTypePicker } from "../EvidenceTypePicker";
 import type { EvidenceTypeValue } from "../../types/evidence";
 import type { DragScrollController } from "../StepList/dragAutoScroll";
-import { ConfirmDeleteModal } from "../ConfirmDeleteModal";
-import { EditGoalStepRow } from "./EditGoalStepRow";
-import { EditGoalSubStepList } from "./EditGoalSubStepList";
-import { useEditGoalDrag } from "./useEditGoalDrag";
+import { EditGoalStepList } from "./EditGoalStepList";
 import { styles } from "./EditGoalView.styles";
 
 /** Tone of a display-only date/dependency chip on a step row (D5). */
@@ -194,9 +186,6 @@ export interface EditGoalViewProps {
   deleteSubStepConfirmMessage?: (title: string) => string;
 }
 
-const defaultStepCountLabel = (count: number) =>
-  `${count} ${count === 1 ? "step" : "steps"}`;
-
 export function EditGoalView({
   goalTitle,
   onGoalTitleChange,
@@ -219,227 +208,33 @@ export function EditGoalView({
   dragScrollController,
   headerLabel = "Edit goal",
   goalSectionLabel = "Goal",
-  stepsSectionLabel = "Steps",
-  addStepPlaceholder = "Add step...",
   descriptionPlaceholder,
   descriptionSectionLabel = "Goal description",
   datesInfoText = 'Dates & dependencies live on each step — tap a step in the full planner to set "after" / "waiting on".',
   doneLabel = "Done",
   overflowAccessibilityLabel = "More options",
-  evidencePickerTitle = "Planned evidence",
-  evidenceTypesLabel = "Evidence types",
-  stepCountLabel = defaultStepCountLabel,
-  addSubStepLabel = "add a sub-step",
-  breakIntoSubStepsLabel = "break into sub-steps",
-  newSubStepTitle = "New sub-step",
-  addStepButtonLabel = "Add step",
-  closeLabel = "Close",
-  breakIntoSubStepsA11yLabel = (stepTitle) =>
-    `Break "${stepTitle}" into sub-steps`,
-  addSubStepA11yLabel = (stepTitle) => `Add a sub-step to "${stepTitle}"`,
+  // Copy consumed only by the step-row list layer — forwarded straight to
+  // EditGoalStepList, which owns their English defaults (D2). Not defaulted
+  // here so there's a single source of truth per prop.
+  stepsSectionLabel,
+  addStepPlaceholder,
+  evidencePickerTitle,
+  evidenceTypesLabel,
+  stepCountLabel,
+  addSubStepLabel,
+  breakIntoSubStepsLabel,
+  newSubStepTitle,
+  addStepButtonLabel,
+  closeLabel,
+  breakIntoSubStepsA11yLabel,
+  addSubStepA11yLabel,
   announceReorder,
-  deleteStepConfirmTitle = "Delete step?",
-  deleteStepConfirmMessage = (title) =>
-    `Delete "${title}"? Its evidence and any sub-steps will be removed too.`,
-  deleteSubStepConfirmTitle = "Delete sub-step?",
-  deleteSubStepConfirmMessage = (title) =>
-    `Delete "${title}"? Its evidence will be removed too.`,
+  deleteStepConfirmTitle,
+  deleteStepConfirmMessage,
+  deleteSubStepConfirmTitle,
+  deleteSubStepConfirmMessage,
 }: EditGoalViewProps) {
   const { theme } = useUnistyles();
-  const { animationPref } = useAnimationPref();
-
-  const [newStepTitle, setNewStepTitle] = useState("");
-  // A single "which id is being renamed" source, keyed by step OR sub-step id
-  // (ids are unique across both). commitEditing routes to the right callback.
-  const [editingId, setEditingId] = useState<string | null>(null);
-  const [editText, setEditText] = useState("");
-  // Which step or sub-step's evidence picker is open (D8/D12). Internal state —
-  // the chip tap opens it locally; onStepEvidenceChange / onSubStepEvidenceChange
-  // are the only outward evidence events.
-  const [editingEvidenceId, setEditingEvidenceId] = useState<string | null>(
-    null,
-  );
-  const [screenReaderActive, setScreenReaderActive] = useState(false);
-  // Which row-level delete is awaiting confirmation (#460, D1/D2). One modal
-  // instance below is driven by this; the rows only signal intent.
-  const [pendingDelete, setPendingDelete] = useState<{
-    kind: "step" | "subStep";
-    id: string;
-    title: string;
-  } | null>(null);
-
-  const drag = useEditGoalDrag({
-    steps,
-    onReorderSteps,
-    dragScrollController,
-    announceReorder,
-  });
-
-  useEffect(() => {
-    AccessibilityInfo.isScreenReaderEnabled()
-      .then(setScreenReaderActive)
-      .catch(() => {
-        // Fail open: show the ↑/↓ fallback if we can't determine SR status.
-        setScreenReaderActive(true);
-      });
-    const sub = AccessibilityInfo.addEventListener(
-      "screenReaderChanged",
-      setScreenReaderActive,
-    );
-    return () => sub.remove();
-  }, []);
-
-  const showAccessibleControls = screenReaderActive || animationPref === "none";
-  const canDrag = steps.length > 1 && editingId === null;
-
-  function handleAddStep() {
-    const trimmed = newStepTitle.trim();
-    if (!trimmed) return;
-    onAddStep(trimmed);
-    setNewStepTitle("");
-  }
-
-  // Find the sub-step with `id` across every parent (one-level, D12).
-  function findSubStep(id: string): EditGoalSubStep | undefined {
-    for (const s of steps) {
-      const sub = s.subSteps?.find((ss) => ss.id === id);
-      if (sub) return sub;
-    }
-    return undefined;
-  }
-
-  function beginEdit(id: string, title: string) {
-    setEditingId(id);
-    setEditText(title);
-  }
-
-  function commitEditing() {
-    if (editingId) {
-      const trimmed = editText.trim();
-      const step = steps.find((s) => s.id === editingId);
-      if (step) {
-        if (trimmed && trimmed !== step.title) {
-          onStepTitleChange(editingId, trimmed);
-        }
-      } else {
-        const sub = findSubStep(editingId);
-        if (sub && trimmed && trimmed !== sub.title) {
-          onSubStepTitleChange(editingId, trimmed);
-        }
-      }
-    }
-    setEditingId(null);
-    setEditText("");
-  }
-
-  function handleAddSubStep(parentStepId: string) {
-    onAddSubStep(parentStepId, newSubStepTitle);
-  }
-
-  // Evidence-picker toggle (D8/D12). Guards the "every step requires evidence"
-  // invariant: the last remaining type can't be deselected (no-op), so a step or
-  // sub-step never lands in a 0-selected state. Routes to the step or sub-step
-  // callback depending on which id opened the picker.
-  function handleToggleEvidence(type: EvidenceTypeValue) {
-    if (editingEvidenceId === null) return;
-    const step = steps.find((s) => s.id === editingEvidenceId);
-    const sub = step ? undefined : findSubStep(editingEvidenceId);
-    const current = step?.plannedEvidenceTypes ?? sub?.plannedEvidenceTypes;
-    if (!current) return;
-    const isSelected = current.includes(type);
-    if (isSelected && current.length === 1) return;
-    const next = isSelected
-      ? current.filter((t) => t !== type)
-      : [...current, type];
-    if (step) {
-      onStepEvidenceChange(editingEvidenceId, next);
-    } else {
-      onSubStepEvidenceChange(editingEvidenceId, next);
-    }
-  }
-
-  const editingEvidenceStep =
-    editingEvidenceId !== null
-      ? steps.find((s) => s.id === editingEvidenceId)
-      : undefined;
-  const editingEvidenceSub =
-    editingEvidenceId !== null && !editingEvidenceStep
-      ? findSubStep(editingEvidenceId)
-      : undefined;
-  const editingEvidenceTypes =
-    editingEvidenceStep?.plannedEvidenceTypes ??
-    editingEvidenceSub?.plannedEvidenceTypes;
-
-  // Sub-steps block (D12), rendered inside each parent's card. A parent with
-  // no sub-steps shows the "break into sub-steps" prompt; one with some
-  // shows the indented green-rail block of sub-rows plus "add a sub-step".
-  // Both add affordances seed a default-titled sub-step (renamed inline after).
-  function renderSubStepBlock(step: EditGoalStep) {
-    const subs = step.subSteps ?? [];
-    if (subs.length === 0) {
-      return (
-        <Pressable
-          style={styles.breakIntoRow}
-          onPress={() => handleAddSubStep(step.id)}
-          accessibilityRole="button"
-          accessibilityLabel={breakIntoSubStepsA11yLabel(step.title)}
-          hitSlop={6}
-          testID={`edit-goal-break-into-${step.id}`}
-        >
-          <RNText
-            style={styles.breakIntoGlyph}
-            accessibilityElementsHidden
-            importantForAccessibility="no"
-          >
-            {"⚊"}
-          </RNText>
-          <RNText style={styles.breakIntoText}>{breakIntoSubStepsLabel}</RNText>
-        </Pressable>
-      );
-    }
-    return (
-      <View style={styles.subStepBlock}>
-        <EditGoalSubStepList
-          subSteps={subs}
-          onReorder={(ids) => onReorderSubSteps(step.id, ids)}
-          editingId={editingId}
-          editText={editText}
-          onEditTextChange={setEditText}
-          onStartEditing={(id, title) => beginEdit(id, title)}
-          onCommitEditing={commitEditing}
-          onEvidenceChipPress={(id) => setEditingEvidenceId(id)}
-          onDelete={(id) =>
-            setPendingDelete({
-              kind: "subStep",
-              id,
-              title: subs.find((s) => s.id === id)?.title ?? "",
-            })
-          }
-          showAccessibleControls={showAccessibleControls}
-          animationPref={animationPref}
-          dragScrollController={dragScrollController}
-          announceReorder={announceReorder}
-        />
-        <Pressable
-          style={styles.addSubStepRow}
-          onPress={() => handleAddSubStep(step.id)}
-          accessibilityRole="button"
-          accessibilityLabel={addSubStepA11yLabel(step.title)}
-          hitSlop={6}
-          testID={`edit-goal-add-substep-${step.id}`}
-        >
-          <RNText
-            style={styles.addSubStepGlyph}
-            accessibilityElementsHidden
-            importantForAccessibility="no"
-          >
-            {"+"}
-          </RNText>
-          <RNText style={styles.addSubStepText}>{addSubStepLabel}</RNText>
-        </Pressable>
-      </View>
-    );
-  }
 
   return (
     <View style={styles.container}>
@@ -489,105 +284,40 @@ export function EditGoalView({
           <Pencil size={16} weight="bold" color={theme.colors.textSecondary} />
         </View>
 
-        <View style={styles.stepsHeader}>
-          <Text
-            variant="headline"
-            style={styles.stepsLabel}
-            accessibilityRole="header"
-          >
-            {stepsSectionLabel}
-          </Text>
-          <RNText style={styles.stepCount}>
-            {stepCountLabel(steps.length)}
-          </RNText>
-        </View>
-
-        <GestureHandlerRootView style={styles.stepList}>
-          {steps.map((step, index) => (
-            <View
-              key={step.id}
-              onLayout={(e) =>
-                drag.registerRowLayout(index, {
-                  y: e.nativeEvent.layout.y,
-                  height: e.nativeEvent.layout.height,
-                })
-              }
-            >
-              <EditGoalStepRow
-                step={step}
-                index={index}
-                stepNumber={index + 1}
-                isBeingDragged={drag.draggedIndex === index}
-                isEditing={editingId === step.id}
-                editText={editText}
-                onEditTextChange={setEditText}
-                onStartEditing={() => beginEdit(step.id, step.title)}
-                onCommitEditing={commitEditing}
-                onEvidenceChipPress={() => setEditingEvidenceId(step.id)}
-                onDragStart={drag.handleDragStart}
-                onDragMove={drag.handleDragMove}
-                onDragEnd={drag.handleDragEnd}
-                dragScrollCompensation={
-                  drag.draggedIndex === index
-                    ? drag.dragScrollCompensation
-                    : undefined
-                }
-                onMoveUp={() => drag.moveStep(index, -1)}
-                onMoveDown={() => drag.moveStep(index, 1)}
-                showAccessibleControls={showAccessibleControls}
-                animationPref={animationPref}
-                isFirst={index === 0}
-                isLast={index === steps.length - 1}
-                canDrag={canDrag}
-                onDelete={() =>
-                  setPendingDelete({
-                    kind: "step",
-                    id: step.id,
-                    title: step.title,
-                  })
-                }
-              >
-                {/* Sub-steps block (D12), rendered inside the parent card so
-                    it drags with the parent. See renderSubStepBlock. */}
-                {renderSubStepBlock(step)}
-              </EditGoalStepRow>
-            </View>
-          ))}
-          {drag.isDragging && drag.dropSlot && (
-            <View
-              pointerEvents="none"
-              accessibilityElementsHidden
-              importantForAccessibility="no"
-              style={[styles.dropLine, { top: drag.dropSlot.top }]}
-            />
-          )}
-        </GestureHandlerRootView>
-
-        <View style={styles.addRow}>
-          <View style={styles.addInputCard}>
-            <TextInput
-              style={styles.addInput}
-              value={newStepTitle}
-              onChangeText={setNewStepTitle}
-              onSubmitEditing={handleAddStep}
-              placeholder={addStepPlaceholder}
-              placeholderTextColor={theme.colors.textMuted}
-              returnKeyType="done"
-              blurOnSubmit={false}
-              testID="edit-goal-add-step-input"
-              accessibilityLabel={addStepPlaceholder}
-            />
-          </View>
-          <Pressable
-            style={styles.addButton}
-            onPress={handleAddStep}
-            testID="edit-goal-add-step-button"
-            accessibilityRole="button"
-            accessibilityLabel={addStepButtonLabel}
-          >
-            <RNText style={styles.addButtonText}>+</RNText>
-          </Pressable>
-        </View>
+        {/* Step-row list layer (#489): "Steps" header + count, drag-reorderable
+            rows, sub-step blocks, add-step affordance, and the evidence-picker /
+            confirm-delete modals. All editing state lives inside it. */}
+        <EditGoalStepList
+          steps={steps}
+          onReorderSteps={onReorderSteps}
+          onReorderSubSteps={onReorderSubSteps}
+          onAddStep={onAddStep}
+          onStepTitleChange={onStepTitleChange}
+          onStepEvidenceChange={onStepEvidenceChange}
+          onAddSubStep={onAddSubStep}
+          onSubStepTitleChange={onSubStepTitleChange}
+          onSubStepEvidenceChange={onSubStepEvidenceChange}
+          onDeleteSubStep={onDeleteSubStep}
+          onDeleteStep={onDeleteStep}
+          dragScrollController={dragScrollController}
+          stepsSectionLabel={stepsSectionLabel}
+          addStepPlaceholder={addStepPlaceholder}
+          evidencePickerTitle={evidencePickerTitle}
+          evidenceTypesLabel={evidenceTypesLabel}
+          stepCountLabel={stepCountLabel}
+          addSubStepLabel={addSubStepLabel}
+          breakIntoSubStepsLabel={breakIntoSubStepsLabel}
+          newSubStepTitle={newSubStepTitle}
+          addStepButtonLabel={addStepButtonLabel}
+          closeLabel={closeLabel}
+          breakIntoSubStepsA11yLabel={breakIntoSubStepsA11yLabel}
+          addSubStepA11yLabel={addSubStepA11yLabel}
+          announceReorder={announceReorder}
+          deleteStepConfirmTitle={deleteStepConfirmTitle}
+          deleteStepConfirmMessage={deleteStepConfirmMessage}
+          deleteSubStepConfirmTitle={deleteSubStepConfirmTitle}
+          deleteSubStepConfirmMessage={deleteSubStepConfirmMessage}
+        />
 
         <View style={styles.infoBanner}>
           <RNText
@@ -609,85 +339,6 @@ export function EditGoalView({
           testID="edit-goal-done-button"
         />
       </View>
-
-      {/* Evidence-type picker (D8/D12): the reused multi-select authoring grid
-          in a local bottom-sheet Modal (mirrors the capture branch + nest
-          picker). Opened by a step OR a sub-step's chip; toggling updates that
-          row's pills via onStepEvidenceChange / onSubStepEvidenceChange; the last
-          remaining type can't be deselected (handleToggleEvidence guard). */}
-      <Modal
-        visible={editingEvidenceTypes !== undefined}
-        transparent
-        animationType="slide"
-        onRequestClose={() => setEditingEvidenceId(null)}
-        accessibilityViewIsModal
-      >
-        <View style={styles.pickerOverlay}>
-          <Pressable
-            style={styles.pickerBackdrop}
-            onPress={() => setEditingEvidenceId(null)}
-            accessibilityRole="button"
-            accessibilityLabel={closeLabel}
-            testID="edit-goal-evidence-backdrop"
-          />
-          <SafeAreaView edges={["bottom"]} style={styles.pickerSheet}>
-            <View style={styles.pickerHandle} />
-            <View style={styles.pickerHeader}>
-              <RNText style={styles.pickerTitle} accessibilityRole="header">
-                {evidencePickerTitle}
-              </RNText>
-              <Pressable
-                style={styles.pickerClose}
-                onPress={() => setEditingEvidenceId(null)}
-                accessibilityRole="button"
-                accessibilityLabel={closeLabel}
-                hitSlop={8}
-                testID="edit-goal-evidence-close"
-              >
-                <RNText style={styles.pickerCloseIcon}>{"✕"}</RNText>
-              </Pressable>
-            </View>
-            {editingEvidenceTypes ? (
-              <EvidenceTypePicker
-                selectedTypes={editingEvidenceTypes}
-                onToggleType={handleToggleEvidence}
-                label={evidenceTypesLabel}
-              />
-            ) : null}
-          </SafeAreaView>
-        </View>
-      </Modal>
-
-      {/* Confirm-delete modal (#460, D1/D2): one instance for both row-level
-          deletions. The main-step × and the sub-step × both open it via
-          pendingDelete; onDeleteStep / onDeleteSubStep fire only on Confirm.
-          The modal supplies its own themed, i18n-aware Delete/Cancel buttons. */}
-      <ConfirmDeleteModal
-        visible={pendingDelete !== null}
-        onCancel={() => setPendingDelete(null)}
-        onConfirm={() => {
-          if (pendingDelete) {
-            if (pendingDelete.kind === "step") {
-              onDeleteStep(pendingDelete.id);
-            } else {
-              onDeleteSubStep(pendingDelete.id);
-            }
-          }
-          setPendingDelete(null);
-        }}
-        title={
-          pendingDelete?.kind === "subStep"
-            ? deleteSubStepConfirmTitle
-            : deleteStepConfirmTitle
-        }
-        message={
-          pendingDelete
-            ? pendingDelete.kind === "subStep"
-              ? deleteSubStepConfirmMessage(pendingDelete.title)
-              : deleteStepConfirmMessage(pendingDelete.title)
-            : ""
-        }
-      />
     </View>
   );
 }
